+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Display error if username in login form contains invalid characters.
+- Scroll to the top of notification list after navigation.
 
 ### Removed