+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Add tooltip to display name in post block.
+
 ## [2.16.0] - 2024-04-16
 
 ### Added