+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [2.17.1] - 2024-04-29
+
 ### Changed
 
 - Hide mention suggestions in preview mode.